+++ wappstoiot/modules/network.py
@@ -27,7 +27,7 @@ class Network(object):
         self,
         name: str,
         connection: ServiceClass,
-        network_uuid: UUID4,
+        network_uuid: uuid.UUID,
         description: str = "",
     ) -> None:
         """
@@ -36,17 +36,14 @@ class Network(object):
         self.log = logging.getLogger(__name__)
         self.log.addHandler(logging.NullHandler())
 
-        # self.closed = False
-
         kwargs = locals()
-        self.__uuid: UUID4 = network_uuid
+        self.__uuid: uuid.UUID = network_uuid
         self.element: WSchema.Network = self.schema()
 
-        self.children_uuid_mapping: Dict[UUID4, Device] = {}
-        # self.children_id_mapping: Dict[int, UUID4] = {}
-        self.children_name_mapping: Dict[str, UUID4] = {}
+        self.children_uuid_mapping: Dict[uuid.UUID, Device] = {}
+        self.children_name_mapping: Dict[str, uuid.UUID] = {}
 
-        self.cloud_id_mapping: Dict[int, UUID4] = {}
+        self.cloud_id_mapping: Dict[int, uuid.UUID] = {}
 
         self.connection: ServiceClass = connection
 
@@ -94,7 +91,7 @@ class Network(object):
         return self.element.name
 
     @property
-    def uuid(self) -> UUID4:
+    def uuid(self) -> uuid.UUID:
         """Returns the name of the value."""
         return self.__uuid
 
@@ -131,8 +128,12 @@ class Network(object):
         # UNSURE(MBK): How should it get the data back?
         """
         def _cb(obj, method):
-            if method == WappstoMethod.PUT:
-                callback(...)
+            try:
+                if method == WappstoMethod.PUT:
+                    callback(...)
+            except Exception:
+                self.log.exception("OnChange callback error.")
+                raise
 
         self.connection.subscribe_network_event(
             uuid=self.uuid,
@@ -147,8 +148,12 @@ class Network(object):
         Configure a callback for when a create have been make for the Device.
         """
         def _cb(obj, method):
-            if method == WappstoMethod.POST:
-                callback()
+            try:
+                if method == WappstoMethod.POST:
+                    callback()
+            except Exception:
+                self.log.exception("onCreate callback error.")
+                raise
 
         self.connection.subscribe_network_event(
             uuid=self.uuid,
@@ -167,8 +172,12 @@ class Network(object):
         # It can not! there is no '{"status":"update"}' that can be set.
         """
         def _cb(obj, method):
-            if method == WappstoMethod.GET:
-                callback()
+            try:
+                if method == WappstoMethod.GET:
+                    callback()
+            except Exception:
+                self.log.exception("onRefresh callback error.")
+                raise
 
         self.connection.subscribe_network_event(
             uuid=self.uuid,
@@ -188,8 +197,12 @@ class Network(object):
         recreated, and/or the program have to close.
         """
         def _cb(obj, method):
-            if method == WappstoMethod.DELETE:
-                callback(self)
+            try:
+                if method == WappstoMethod.DELETE:
+                    callback(self)
+            except Exception:
+                self.log.exception("onDelete callback error.")
+                raise
 
         self.connection.subscribe_network_event(
             uuid=self.uuid,
@@ -254,12 +267,11 @@ class Network(object):
             name=name
         )
 
-        if not device_uuid:
-            kwargs['device_uuid'] = uuid.uuid4()
-        else:
-            kwargs['device_uuid'] = device_uuid[0]
-
-        device_obj = Device(parent=self, **kwargs)
+        device_obj = Device(
+            parent=self,
+            device_uuid=device_uuid,
+            **kwargs
+        )
         self.__add_device(device_obj, kwargs['name'])
         return device_obj